requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
dataclasses-json==0.6.7
deepgram-sdk==2.12.0
deprecation==2.1.0
distro==1.9.0
elevenlabs==1.57.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.76.0
packaging==25.0
pluggy==1.5.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
setuptools==79.0.1
sniffio==1.3.1
sounddevice==0.5.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
webrtcvad==2.0.10
websockets==15.0
yarl==1.20.0
